requirements/wordcloud-requirements.txt
pylint==2.3.1
pandas==0.24.2
numpy==1.16.2
matplotlib==3.0.3
pillow==5.1.0
wordcloud==1.5.0